+++ docs/python/runtime/python.mdx
@@ -1,18 +1,24 @@
 ---
 title: Python
-description: Choose the interpreter that runs python3 inside the workspace, monty (sandboxed, default), wasi (sandboxed full CPython), or the host interpreter.
+description: Choose the interpreter that runs python3 inside the workspace, monty (sandboxed, default), wasi (sandboxed full CPython), sandlock (the host interpreter, confined), or the host interpreter.
 icon: python
 ---
 
 Shell lines like `python3 script.py` need an interpreter. Mirage calls that
-interpreter a **runtime**, and you pick it per workspace. Three runtimes
+interpreter a **runtime**, and you pick it per workspace. Four runtimes
 ship today:
 
 | Runtime | Engine | Filesystem | Default |
 | --- | --- | --- | --- |
 | `monty` | [Pydantic Monty](https://github.com/pydantic/monty), a sandboxed Python interpreter written in Rust | Workspace mounts only, bridged through Mirage | Yes |
 | `wasi` | Full CPython compiled to WebAssembly, run in-process on [wasmtime](https://wasmtime.dev) | Workspace mounts only, bridged through Mirage; no host filesystem, no network | No |
+| `sandlock` | The host `python` interpreter, confined by [sandlock](https://github.com/multikernel/sandlock) (Landlock + seccomp). Linux only | Only the host paths the config grants | No |
 | `local` | The host `python` interpreter, as a subprocess | The host filesystem | No |
+
+The first two bridge file I/O through the workspace, so code sees your
+mounts and nothing else. The last two run real CPython on the host and see
+the host filesystem instead — `sandlock` bounded to the paths you name,
+`local` unbounded.
 
 ## Monty (default)
 
@@ -152,6 +158,54 @@ runtimes:
 
 Note that `local` code does **not** see workspace mounts; it sees the real
 host filesystem.
+
+## <img noZoom src="/images/sandlock-logo.svg" alt="" className="inline-block h-5 align-middle mr-2" />Sandlock
+
+`sandlock` is `local` with a fence around it: the same host CPython, the
+same native wheels and exact `sys.flags`, but spawned through
+[sandlock](https://github.com/multikernel/sandlock) so a Landlock ruleset
+and a seccomp filter bound where its I/O can land. Nothing is granted
+except the interpreter's own tree, the system paths CPython needs to
+start, and what you list:
+
+```yaml
+runtimes:
+  - name: sandlock
+    config:
+      fs_readable: ["/opt/data"]
+      fs_writable: ["/mnt/workspace"]
+      max_memory: "512M"
+  - vfs
+```
+
+It needs the sandlock CLI on `PATH` — there is no extra to install,
+because the CLI is the transport. **Linux only**, and the full ruleset
+wants Landlock ABI v6 (Linux 6.12+); sandlock's own degrade options cover
+older kernels with weaker protection.
+
+Two things to know before reaching for it:
+
+- **It does not see workspace mounts either.** Its reach is `process`,
+  the same as `local`: confinement narrows *which* host paths the code
+  can touch, but the workspace gate never sees that I/O, so mount modes,
+  policy, and accounting do not apply. What it buys is that a `process`
+  runtime need no longer be an unbounded one. Point `fs_writable` at a
+  [FUSE-mounted](/python/setup/fuse) workspace path and the code works on
+  your mounts while reaching nothing else on the host.
+- **The environment is not inherited.** `local` passes `os.environ`
+  straight through; `sandlock` starts empty and sets only the config
+  `env` plus the run's own, so confined code is never handed the backend
+  credentials mirage is holding. Anything the code needs must be named:
+
+  ```yaml
+  runtimes:
+    - name: sandlock
+      config:
+        env:
+          HOME: /tmp
+          TZ: UTC
+    - vfs
+  ```
 
 ## Selecting in YAML
